% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{add_matrix_to_mudata}
\alias{add_matrix_to_mudata}
\title{Add a matrix as a new experiment to existing MuData object}
\usage{
add_matrix_to_mudata(new_matrix, mudata, experiment_name)
}
\arguments{
\item{new_matrix}{A matrix to be added to the MuData object}

\item{mudata}{An existing MuData object}

\item{experiment_name}{A string indicating the name of the new experiment}
}
\value{
A MuData object with the new experiment added
}
\description{
Add a matrix as a new experiment to existing MuData object
}
\examples{
library(sceptreIGVF)
library(MultiAssayExperiment)
# load sample MuData
data(sample_mudata)
# create new matrix
cell_barcodes <- sample_mudata[["scRNA"]] |> colnames()
num_cells <- length(cell_barcodes)
num_features <- 100
new_matrix <- matrix(
  rnorm(num_features * num_cells),
  nrow = num_features, ncol = num_cells,
  dimnames = list(NULL, cell_barcodes)
)
# add new matrix to MuData object
mudata <- add_matrix_to_mudata(new_matrix, sample_mudata, "new_experiment")
mudata
}
