% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grna-assignment.R
\name{assign_grnas_sceptre}
\alias{assign_grnas_sceptre}
\title{Assign gRNAs to cells via sceptre's mixture method}
\usage{
assign_grnas_sceptre(mudata)
}
\arguments{
\item{mudata}{A MuData object}
}
\value{
A MuData object with sceptre gRNA assignments added to the \code{guide} modality
as a new assay called "guide_assignment"
}
\description{
Assign gRNAs to cells via sceptre's mixture method
}
\examples{
library(sceptreIGVF)
# load sample MuData
data(mudata_guide_assignment_gasperini)
# assign gRNAs
mudata_out <- assign_grnas_sceptre(mudata_guide_assignment_gasperini)
mudata_out
}
