% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grna-assignment.R
\name{extract_grna_assignment_matrix}
\alias{extract_grna_assignment_matrix}
\title{Extra gRNA assignment matrix from \code{sceptre} object}
\usage{
extract_grna_assignment_matrix(sceptre_object)
}
\arguments{
\item{sceptre_object}{A sceptre_object}
}
\value{
A sparse logical matrix of gRNA assignments
}
\description{
Extra gRNA assignment matrix from \code{sceptre} object
}
