% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mudata.R
\docType{data}
\name{mudata_guide_assignment}
\alias{mudata_guide_assignment}
\title{Sample MuData}
\format{
An object of class \code{MultiAssayExperiment} of length 2.
}
\usage{
data(mudata_guide_assignment)
}
\description{
A subset of the MuData object from the Gasperini et al. (2019) pilot data
}
\examples{
data(mudata_guide_assignment)
mudata_guide_assignment
}
\keyword{datasets}
