% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mudata.R
\docType{data}
\name{mudata_guide_assignment_gasperini}
\alias{mudata_guide_assignment_gasperini}
\title{Sample MuData for guide assignment (Gasperini)}
\format{
An object of class MuData
}
\usage{
data(mudata_guide_assignment_gasperini)
}
\description{
A MuData object from a subset of the the Gasperini et al. (2019) data
}
