% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mudata.R
\docType{data}
\name{mudata_guide_assignment_papalexi}
\alias{mudata_guide_assignment_papalexi}
\title{Sample MuData for guide assignment (Papalexi)}
\format{
An object of class MuData
}
\usage{
data(mudata_guide_assignment_papalexi)
}
\description{
A MuData object from a subset of the the Papalexi et al. (2021) data
}
