% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mudata.R
\docType{data}
\name{mudata_inference}
\alias{mudata_inference}
\title{Sample MuData for inference}
\format{
An object of class MuData
}
\usage{
data(mudata_inference)
}
\description{
A MuData object from a subset of the the Gasperini et al. (2019) data
}
