% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mudata.R
\docType{data}
\name{mudata_inference_papalexi}
\alias{mudata_inference_papalexi}
\title{Sample MuData for inference (Papalexi)}
\format{
An object of class MuData
}
\usage{
data(mudata_inference_papalexi)
}
\description{
A MuData object from a subset of the the Gasperini et al. (2021) data
}
