% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mudata.R
\docType{data}
\name{sample_mudata}
\alias{sample_mudata}
\title{Sample MuData}
\format{
An object of class \code{MultiAssayExperiment} of length 2.
}
\usage{
data(sample_mudata)
}
\description{
A subset of the MuData object from the Gasperini et al. (2019) pilot data
}
\examples{
data(sample_mudata)
sample_mudata
}
\keyword{datasets}
