% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{save_mudata_list}
\alias{save_mudata_list}
\title{Save a list of MuData objects to disk}
\usage{
save_mudata_list(mudata_list, path = "./", prefix = "")
}
\arguments{
\item{mudata_list}{List of MuData objects}

\item{path}{Path to save files to}

\item{prefix}{Prefix to add to file names}
}
\description{
Save a list of MuData objects to disk
}
