% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{sceptre_object_to_mudata}
\alias{sceptre_object_to_mudata}
\title{Convert sceptre_object to MuData}
\usage{
sceptre_object_to_mudata(sceptre_object)
}
\arguments{
\item{sceptre_object}{A sceptre_object}
}
\value{
A MuData object
}
\description{
Convert sceptre_object to MuData
}
