% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{sceptre_object_to_mudata_inputs_outputs}
\alias{sceptre_object_to_mudata_inputs_outputs}
\title{Convert sceptre_object a set of MuData objects to be used as inputs and outputs
for both gRNA assignment and inference}
\usage{
sceptre_object_to_mudata_inputs_outputs(
  sceptre_object,
  num_discovery_pairs,
  gene_info = NULL,
  guide_capture_method = NULL
)
}
\arguments{
\item{sceptre_object}{Sceptre object}

\item{num_discovery_pairs}{Number of discovery pairs to sample}

\item{gene_info}{Gene information data frame (optional)}

\item{guide_capture_method}{Guide capture method (optional)}
}
\value{
List of MuData objects
}
\description{
Convert sceptre_object a set of MuData objects to be used as inputs and outputs
for both gRNA assignment and inference
}
