% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{convert_mudata_to_sceptre_object}
\alias{convert_mudata_to_sceptre_object}
\title{Convert IGVF CRISPR MuData to sceptre_object}
\usage{
convert_mudata_to_sceptre_object(mudata)
}
\arguments{
\item{mudata}{IGVF CRISPR MuData object}
}
\value{
A sceptre_object
}
\description{
Convert IGVF CRISPR MuData to sceptre_object
}
\examples{
library(sceptreIGVF)
# load sample MuData for guide assignment
data(mudata_guide_assignment)
# convert MuData object to sceptre object
sceptre_object_guide_assignment <- convert_mudata_to_sceptre_object(mudata_guide_assignment)
sceptre_object_guide_assignment
# load sample MuData for inference
data(mudata_inference)
# convert MuData object to sceptre object
sceptre_object_inference <- convert_mudata_to_sceptre_object(mudata_inference)
sceptre_object_inference
}
