% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mudata.R
\name{convert_mudata_to_sceptre_object}
\alias{convert_mudata_to_sceptre_object}
\title{Convert IGVF CRISPR MuData to sceptre_object}
\usage{
convert_mudata_to_sceptre_object(mudata)
}
\arguments{
\item{mudata}{IGVF CRISPR MuData object}

\item{moi}{A string indicating the multiplicity of infection ("high" or "low")}

\item{include_covariates}{A logical indicating whether to include the
covariates from colData in the sceptre object as extra covariates}
}
\value{
A sceptre_object
}
\description{
Convert IGVF CRISPR MuData to sceptre_object
}
\examples{
library(sceptreIGVF)
# load sample MuData
data(sample_mudata)
# convert MuData object to sceptre object
sceptre_object <- convert_mudata_to_sceptre_object(sample_mudata)
sceptre_object
}
