% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_sceptre}
\alias{inference_sceptre}
\title{SCEPTRE inference}
\usage{
inference_sceptre(mudata, ...)
}
\arguments{
\item{mudata}{A MuData object}

\item{...}{Additional arguments passed to sceptre::set_analysis_parameters();
see \code{?sceptre::set_analysis_parameters} for details.}
}
\value{
A MuData object with sceptre inference results
}
\description{
SCEPTRE inference
}
\examples{
library(sceptreIGVF)
# load sample MuData
data(mudata_inference_gasperini)
# inference
mudata_out <- inference_sceptre(mudata_inference_gasperini)
mudata_out
}
